[fix] catch and retry batches if we get a bad token
It's possible for the sequence tokens to get out of whack if multiple
process write to the log stream. We avoided this by having each pod
write to it's own log stream, but this still comes up on occassion.

This fix will retry the batch if we encounter the invalid token error
and reset the sequence token to the expected sequence token.

RHCLOUD-12772

Signed-off-by: Stephen Adams <tsadams@gmail.com>